Rewrite build tools to parse profiles based on CaC control files
Previous implementation parsed CaC profile files,
which did not rely on controls files.
All the rules had to be defined as selections in the profile,
preceeded by comments containing CIS IDs and titles.

This implementation still parses the CaC profile files,
but extracts the rules, variables and control metadata (id, title)
from the control files, if defined. The rules defined
or overriden in the profile are also considered, however,
any comments containing metadata are now ignored.