Improve id of 'profile' control in build tool
The 'profile' control is a fake control used by the build tool to
store rules and vars which don't belong to any control in the controls
file, but which are defined/overriden in the actual profile.

This change improves the clarity of logs and tailoring file comments
by changing the id from ProfileControl to Profile_Rules_And_Variables.