Introduce state file to retain the benchmark version
The version of the benchmark profile used in usg commands
is now set during first execution of usg and retained
until explicitly redefined using the -V/--benchmark-version flag.

During the initial usg run, or if the state file is removed,
or if using `--benchmark-version "latest"`, the latest version
of the specified benchmark profile is selected and stored
in the state file /var/lib/usg/state.json.

This makes the process more aligned with tailoring files which
are also pinned to a specific version to avoid an unwanted automatic
upgrade to a backwards incompatible version of the benchmark.