Improve profile search and error messaging
Existing get_profile() implementation only checks if the combination of
profile/product/version exists.

This is replaced with an iterative approach, which informs the user
explicitly whether the profile name cannot be found in any benchmark,
or whether the product name cannot be found among the matching profiles,
or whether there is no version that corresponds to that profile name.

Additionally, a fallback was added to the search when requesting the
latest version of the profile.  If the latest version of the benchmark
does not contain the said profile, the most recent major version of
the benchmark (containing the correct profile) is chosen,
based on the tailoring version.